move grantslist to index
