+++ src/interfaces/Grant.ts
@@ -1,6 +1,7 @@
 export interface Grant {
   id: number;
   expiryDate: string;
+  expiryDateFormatted?: string;
   title: string;
   description: string;
   submittedBy: string;
